rag-api/requirements.txt
python-dotenv # For reading environment variables stored in .env file
langchain
langchain-community
langchain-openai # For embeddings
unstructured # Document loading
# onnxruntime1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb # Vector storage
openai # For embeddings
tiktoken  # For embeddings 

fastapi
uvicorn

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 
